eyeson/backend/requirements.txt
# EYESON Backend Requirements
# Enterprise BodyScan Platform

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.6

# Configuration
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-jwt==4.1.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Task Queue
celery==5.3.6
flower==2.0.1

# Open Source TTS
# Kokoro-82M uses ONNX Runtime
onnxruntime==1.17.0
# Piper TTS (optional fallback)
piper-tts==1.2.0

# STT (Deepgram)
deepgram-sdk==3.1.0

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
scipy==1.12.0
numpy==1.26.3

# Video Processing
opencv-python==4.9.0.80
ffmpeg-python==0.2.0

# Image Processing
Pillow==10.2.0

# ML/AI
# Note: PyTorch should be installed separately based on CUDA version
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
torch==2.2.0
torchvision==0.17.0
transformers==4.37.2
diffusers==0.26.2
accelerate==0.27.0

# 3D Processing
trimesh==4.0.9
pyrender==0.1.45

# Cloud Storage
boto3==1.34.34
botocore==1.34.34

# Monitoring & Logging
prometheus-client==0.19.0
structlog==24.1.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0
factory-boy==3.3.0

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# Utilities
python-dateutil==2.8.2
tenacity==8.2.3
orjson==3.9.12

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.7
